+++ db.json
@@ -5,7 +5,7 @@
       "alertMessage": "message",
       "activeAlarm": "./singingBowl.mp3",
       "volume": 34,
-      "time": 270,
+      "time": 150,
       "alert": false
     }
   ]
