+++ src/app/components/App.tsx
@@ -1,6 +1,6 @@
 import Box from '@mui/material/Box';
 import Container  from "@mui/material/Container";
-import Posture   from '../../widgits/components/posture'
+import Posture   from '../../widgits/components/Posture'
 
 function App() {
   return (
